# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Monster Realms Online API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /players/{playerId}:
    get:
      operationId: getPlayer
      tags: [players]
      summary: Get a player profile
      parameters:
        - name: playerId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Player profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
        "404":
          description: Player not found
    patch:
      operationId: updatePlayer
      tags: [players]
      summary: Update player profile
      security:
        - bearerAuth: []
      parameters:
        - name: playerId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlayerUpdate"
      responses:
        "200":
          description: Updated player
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
  /players/{playerId}/stats:
    get:
      operationId: getPlayerStats
      tags: [players]
      summary: Get detailed player statistics
      parameters:
        - name: playerId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Player stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerStats"
  /players/{playerId}/team:
    get:
      operationId: getPlayerTeam
      tags: [players]
      summary: Get the player's active team (up to 6 monsters)
      security:
        - bearerAuth: []
      parameters:
        - name: playerId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Active team
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CapturedMonster"
    put:
      operationId: updatePlayerTeam
      tags: [players]
      summary: Set the player's active team
      security:
        - bearerAuth: []
      parameters:
        - name: playerId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamUpdate"
      responses:
        "200":
          description: Updated team
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CapturedMonster"
  /players/{playerId}/radar:
    get:
      operationId: getPlayerRadar
      tags: [players]
      summary: Get monster radar for the player's current position
      security:
        - bearerAuth: []
      parameters:
        - name: playerId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Radar result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RadarResult"

  # MONSTERS (encyclopedia)
  /monsters:
    get:
      operationId: listMonsterSpecies
      tags: [monsters]
      summary: List all monster species
      parameters:
        - name: element
          in: query
          schema:
            type: string
        - name: rarity
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Monster species list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MonsterSpecies"
  /monsters/{speciesId}:
    get:
      operationId: getMonsterSpecies
      tags: [monsters]
      summary: Get a monster species by ID
      parameters:
        - name: speciesId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Monster species
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonsterSpecies"
        "404":
          description: Not found

  # COLLECTION
...
  /regions/{regionId}:
    get:
      operationId: getRegion
      tags: [regions]
      summary: Get a region by ID
      parameters:
        - name: regionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Region
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Region"
        "404":
          description: Not found

  # EXPLORATION
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Player:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        avatarColor:
          type: string
        isGuest:
          type: boolean
        explorerRank:
          type: string
        explorerXp:
          type: number
        explorerLevel:
          type: number
        tilesExplored:
          type: number
        energy:
          type: number
        maxEnergy:
          type: number
        coins:
          type: number
        monstersDiscovered:
          type: number
        monstersCaptured:
          type: number
        regionId:
          type: ["string", "null"]
        posX:
          type: number
        posY:
          type: number
        createdAt:
          type: string
      required:
        - id
        - username
        - avatarColor
        - isGuest
        - explorerRank
        - explorerXp
        - explorerLevel
        - tilesExplored
        - energy
        - maxEnergy
        - coins
        - monstersDiscovered
        - monstersCaptured
        - posX
        - posY
        - createdAt
    PlayerUpdate:
      type: object
      properties:
        username:
          type: string
          minLength: 2
          maxLength: 20
        avatarColor:
          type: string
    PlayerStats:
      type: object
      properties:
        playerId:
          type: string
        pvpWins:
          type: number
        pvpLosses:
          type: number
        monstersCaptured:
          type: number
        monstersDiscovered:
          type: number
        tilesExplored:
          type: number
        secretsFound:
          type: number
        battlesWon:
          type: number
        battlesLost:
          type: number
        totalPlayTime:
          type: number
        firstDiscoveries:
          type: number
      required:
        - playerId
        - pvpWins
        - pvpLosses
        - monstersCaptured
        - monstersDiscovered
        - tilesExplored
        - secretsFound
        - battlesWon
        - battlesLost
        - totalPlayTime
        - firstDiscoveries
    TeamUpdate:
      type: object
      properties:
        capturedMonsterIds:
          type: array
          items:
            type: string
          maxItems: 6
      required:
        - capturedMonsterIds
    RadarResult:
      type: object
      properties:
        monstersNearby:
          type: boolean
        direction:
          type: ["string", "null"]
          enum: [N, S, E, W, NE, NW, SE, SW, null]
        element:
          type: ["string", "null"]
        distance:
          type: ["number", "null"]
      required:
        - monstersNearby

    # MONSTERS
    MonsterSkill:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [normal, skill1, skill2, ultimate, passive]
        element:
          type: string
        power:
          type: number
        accuracy:
          type: number
        description:
          type: string
      required:
        - name
        - type
        - element
        - power
        - accuracy
        - description
    MonsterSpecies:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        element:
          type: string
          enum: [Fire, Water, Nature, Electric, Ice, Earth, Air, Light, Dark, Metal, Crystal, Void]
        rarity:
          type: string
          enum: [Common, Uncommon, Rare, Epic, Legendary, Mythic, Ancient, Celestial, Void]
        baseHp:
          type: number
        baseAttack:
          type: number
        baseDefense:
          type: number
        baseSpeed:
          type: number
        description:
          type: string
        lore:
          type: string
        captureRate:
          type: number
        weight:
          type: number
        height:
          type: number
        personality:
          type: string
        evolutionFromId:
          type: ["string", "null"]
        evolutionToId:
          type: ["string", "null"]
        evolutionLevel:
          type: ["number", "null"]
        skills:
          type: array
          items:
            $ref: "#/components/schemas/MonsterSkill"
        regionIds:
          type: array
          items:
            type: string
      required:
        - id
        - name
        - element
        - rarity
        - baseHp
        - baseAttack
        - baseDefense
        - baseSpeed
        - description
        - lore
        - captureRate
        - weight
        - height
        - personality
        - skills
        - regionIds

    # COLLECTION
    CapturedMonster:
      type: object
      properties:
        id:
          type: string
        playerId:
          type: string
        speciesId:
          type: string
        species:
          $ref: "#/components/schemas/MonsterSpecies"
        nickname:
          type: ["string", "null"]
        level:
          type: number
        currentHp:
          type: number
        maxHp:
          type: number
        attack:
          type: number
        defense:
          type: number
        speed:
          type: number
        experience:
          type: number
        shinyVariant:
          type: ["string", "null"]
          enum: [Golden, Crystal, Shadow, Galaxy, Prismatic, null]
        friendship:
          type: number
        personality:
          type: string
        inTeam:
          type: boolean
        teamSlot:
          type: ["number", "null"]
        capturedAt:
          type: string
      required:
        - id
        - playerId
        - speciesId
        - species
        - level
        - currentHp
        - maxHp
        - attack
        - defense
        - speed
        - experience
        - friendship
        - personality
        - inTeam
        - capturedAt
...
    Region:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        biome:
          type: string
        description:
          type: string
        requiredExplorerLevel:
          type: number
        monsterCount:
          type: number
        width:
          type: number
        height:
          type: number
        monsters:
          type: array
          items:
            $ref: "#/components/schemas/MonsterSpecies"
      required:
        - id
        - name
        - biome
        - description
        - requiredExplorerLevel
        - monsterCount
        - width
        - height
        - monsters

    # EXPLORATION
...